requirements.txt
streamlit>=1.20.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=9.0.0
scikit-learn>=1.0.0
